update: Firebase SDK のアプデ対応 8.8.1 → 9.6.7(#66)
