+++ package.json
@@ -36,7 +36,7 @@
     "@testing-library/jest-dom": "^5.14.1",
     "@testing-library/react": "^12.0.0",
     "@testing-library/user-event": "^13.2.1",
-    "firebase": "^8.8.1",
+    "firebase": "^9.6.7",
     "konva": "^8.3.2",
     "lodash": "^4.17.21",
     "prop-types": "^15.8.1",
