--- src/libs/Firebase.ts
@@ -1,17 +1,0 @@
-import firebase from 'firebase/app';
-import 'firebase/analytics';
-
-const config = {
-  apiKey: process.env.REACT_APP_APP_KEY,
-  authDomain: process.env.REACT_APP_AUTH_DOMAIN,
-  databaseURL: process.env.REACT_APP_DATABASE_URL,
-  projectId: process.env.REACT_APP_PROJECT_ID,
-  storageBucket: process.env.REACT_APP_STORAGEBUCKET,
-  messagingSenderId: process.env.REACT_APP_MESSAGING_SENDER_ID,
-  appId: process.env.REACT_APP_APP_ID,
-  measurementId: process.env.REACT_APP_MEASUREMENT_ID,
-};
-
-firebase.initializeApp(config);
-
-export default firebase;
